added new webpack configs and webpack5 package ependencies. updated docker files for new build:prd package.json  scripts
