+++ package.json
@@ -4,13 +4,15 @@
   "description": "xForge - Language Forge and Scripture Forge",
   "private": true,
   "scripts": {
+    "build:prd": "npm-run-all -p -l sass webpack:prd",
     "build": "npm-run-all -p -l sass webpack",
     "build:watch": "npm-run-all -r -p -l sass:watch webpack:watch",
     "sass": "node-sass src/sass -o src/dist/css",
     "sass:watch": "npm run sass && node-sass src/sass -o src/dist/css --watch --recursive",
     "webpack": "webpack",
     "webpack:watch": "webpack -w",
-    "webpack:production": "webpack -p"
+    "webpack:production": "webpack -p",
+    "webpack:prd": "webpack --config webpack-prd.config.js"
   },
   "author": "",
   "license": "MIT",
@@ -68,16 +70,20 @@
     "@types/oclazyload": "0.0.32",
     "@types/offline-js": "^0.7.28",
     "@types/rangy": "0.0.30",
+    "@types/webpack-env": "^1.16.0",
     "ajv": "^6.10.0",
     "angular-mocks": "^1.8.2",
-    "awesome-typescript-loader": "^3.2.2",
+    "awesome-typescript-loader": "^5.2.1",
     "babel-core": "^6.26.3",
     "browserstack-local": "^1.4.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.28.11",
+    "copy-webpack-plugin": "^8.1.0",
+    "core-js": "^3.6.5",
+    "coveralls": "^3.1.0",
+    "css-loader": "^3.6.0",
     "del": "^3.0.0",
     "es6-shim": "^0.35.1",
-    "file-loader": "^0.11.2",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "^5.1.0",
     "glob": "^7.1.2",
     "gulp": "^4.0.2",
     "gulp-clean-css": "^4.2.0",
@@ -87,34 +93,48 @@
     "gulp-replace": "^0.5.4",
     "gulp-uglify": "^2.0.0",
     "gulp-util": "~3.0.6",
-    "html-loader": "^1.3.1",
-    "jasmine-core": "^2.7.0",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^5.3.1",
+    "http-server": "^0.12.3",
+    "istanbul-instrumenter-loader": "^3.0.1",
+    "jasmine-core": "^3.6.0",
     "jasmine-reporters": "^2.1.1",
     "jasmine-spec-reporter": "^4.1.1",
     "jshint": "^2.10.2",
-    "karma": "^6.2.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.2",
-    "karma-jasmine": "^1.0.2",
-    "karma-sourcemap-loader": "^0.3.7",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage": "^2.0.3",
+    "karma-jasmine": "^3.3.1",
+    "karma-phantomjs-launcher": "^1.0.4",
+    "karma-sourcemap-loader": "^0.3.8",
     "karma-teamcity-reporter": "^1.0.0",
-    "karma-webpack": "^3.0.0",
+    "karma-webpack": "^4.0.2",
     "lodash.template": "^4.4.0",
     "merge-stream": "^1.0.1",
+    "mini-css-extract-plugin": "^1.3.9",
+    "ng-annotate-loader": "^0.7.0",
     "ngtemplate-loader": "^2.1.0",
     "node-sass": "^4.14.1",
+    "npm-run-all": "^4.1.5",
     "path": "^0.12.7",
     "protractor": "^5.4.2",
     "raw-loader": "^0.5.1",
-    "sass-loader": "^6.0.6",
+    "remap-istanbul": "^0.13.0",
+    "rimraf": "^3.0.2",
+    "sass-loader": "^8.0.2",
     "string-replace-loader": "^1.3.0",
-    "style-loader": "^0.18.2",
-    "tslint": "^5.7.0",
-    "typescript": "^2.5.3",
+    "style-loader": "^1.2.1",
+    "terser-webpack-plugin": "^5.1.1",
+    "tslint": "^5.20.1",
+    "tslint-loader": "^3.6.0",
+    "typescript": "^3.9.7",
+    "uglifyjs-webpack-plugin": "^2.2.0",
     "url-loader": "^0.5.9",
-    "webpack": "^3.4.1",
-    "webpack-livereload-plugin": "^1.2.0",
-    "webpack-merge": "^4.1.0",
+    "webpack": "^5.27.1",
+    "webpack-cli": "^4.5.0",
+    "webpack-dev-server": "^3.11.0",
+    "webpack-livereload-plugin": "^3.0.1",
+    "webpack-merge": "^4.2.2",
     "workbox-webpack-plugin": "^3.3.1",
     "worker-loader": "^2.0.0",
     "yargs": "^16.2.0"
